Improve backward compatibility
The csvread function is now deprecated and does not work well, for example, on NodePosition(i).dat files. The alternative would be readmatrix, but it has been added only from release R2019a.
A work around is to use readtable, introduced in 2013a and able to handle correctly all files it has been tested with, then extracting a matrix from the table.